Refactor package structure and update imports
- Rename types.py to models.py for better organization
- Update all import statements to use correct package naming (mongodb_knowledge_graph)
- Fix package structure to follow Python conventions
- Update dependencies and lock file
- Clean up .gitignore for better project hygiene
