+++ pyproject.toml
@@ -18,7 +18,7 @@ classifiers = [
     "Programming Language :: Python :: 3.12",
 ]
 dependencies = [
-    "mcp>=1.2.0",
+    "mcp[cli]>=1.2.0",
     "pymongo>=4.9.0",
     "pydantic>=2.0.0",
     "typing-extensions>=4.0.0",
@@ -50,7 +50,7 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.wheel]
-packages = ["mcp_server_memory"]
+packages = ["mongodb_knowledge_graph"]
 
 [tool.black]
 line-length = 88
